contracts-common: types: Organize types in module
