+++ contracts-common/src/types/mod.rs
@@ -1,0 +1,14 @@
+//! Types common to all contracts
+
+mod proof_system;
+pub use proof_system::*;
+mod transfers;
+pub use transfers::*;
+mod fees;
+pub use fees::*;
+mod r#match;
+pub use r#match::*;
+mod keys;
+pub use keys::*;
+mod statements;
+pub use statements::*;
